+++ tests/test_hubspot_interrupted_sync.py
@@ -18,7 +18,7 @@ class TestHubspotInterruptedSync1(HubspotBaseTest):
 
     def streams_to_test(self):
         """expected streams minus the streams not under test"""
-        return {'companies', 'engagements'}
+        return {'companies', 'engagements', 'tickets'}
 
     def simulated_interruption(self, reference_state):
 
@@ -37,6 +37,11 @@ class TestHubspotInterruptedSync1(HubspotBaseTest):
         )
         new_state['bookmarks']['engagements']['lastUpdated'] = None
         new_state['bookmarks']['engagements']['current_sync_start'] = engagements_bookmark
+
+        tickets_bookmark = self.timedelta_formatted(
+            reference_state['bookmarks']['tickets']['updatedAt'],
+            days=-1, str_format=self.BASIC_DATE_FORMAT)
+        new_state['bookmarks']['tickets']['updatedAt'] = tickets_bookmark
 
         return new_state